Refactor es_interface trends tests out to separate file
